Update selector to match groups without `field-list`
